# Faster optimized frozen dictionary creation (6/6) 