--- a/src/libraries/Common/tests/System/Collections/ICollection.Generic.Tests.cs
+++ b/src/libraries/Common/tests/System/Collections/ICollection.Generic.Tests.cs
@@ -406,7 +406,7 @@ namespace System.Collections.Tests
         public void ICollection_Generic_Contains_DefaultValueOnCollectionNotContainingDefaultValue(int count)
         {
             ICollection<T> collection = GenericICollectionFactory(count);
-            if (DefaultValueAllowed && default(T) is null) // it's true only for reference types
+            if (DefaultValueAllowed && default(T) is null) // it's true only for reference types and for Nullable<T>
             {
                 Assert.False(collection.Contains(default(T)));
             }