--- a/src/libraries/System.Collections.Immutable/src/System.Collections.Immutable.csproj
+++ b/src/libraries/System.Collections.Immutable/src/System.Collections.Immutable.csproj
@@ -38,6 +38,7 @@ The System.Collections.Immutable library is built-in as part of the shared frame
     <Compile Include="System\Collections\Frozen\Int32\Int32FrozenSet.cs" />
     <Compile Include="System\Collections\Frozen\String\Hashing.cs" />
     <Compile Include="System\Collections\Frozen\String\KeyAnalyzer.cs" />
+    <Compile Include="System\Collections\Frozen\String\LengthBuckets.cs" />
     <Compile Include="System\Collections\Frozen\String\LengthBucketsFrozenDictionary.cs" />
     <Compile Include="System\Collections\Frozen\String\LengthBucketsFrozenSet.cs" />
     <Compile Include="System\Collections\Frozen\String\OrdinalStringFrozenDictionary.cs" />