--- a/src/libraries/System.Collections.Immutable/src/System/Collections/Frozen/Constants.cs
+++ b/src/libraries/System.Collections.Immutable/src/System/Collections/Frozen/Constants.cs
@@ -85,6 +85,6 @@ namespace System.Collections.Frozen
             => typeof(T) == typeof(int) || typeof(T) == typeof(uint)
             || typeof(T) == typeof(short) || typeof(T) == typeof(ushort)
             || typeof(T) == typeof(byte) || typeof(T) == typeof(sbyte)
-            || ((typeof(T) == typeof(IntPtr) || typeof(T) == typeof(UIntPtr)) && IntPtr.Size == 4);
+            || ((typeof(T) == typeof(nint) || typeof(T) == typeof(nuint)) && IntPtr.Size == 4);
     }
 }