--- a/src/libraries/System.Collections.Immutable/src/System/Collections/Frozen/ValueTypeDefaultComparerFrozenSet.cs
+++ b/src/libraries/System.Collections.Immutable/src/System/Collections/Frozen/ValueTypeDefaultComparerFrozenSet.cs
@@ -10,7 +10,7 @@ namespace System.Collections.Frozen
     /// <typeparam name="T">The type of values in the set.</typeparam>
     internal sealed class ValueTypeDefaultComparerFrozenSet<T> : ItemsFrozenSet<T, ValueTypeDefaultComparerFrozenSet<T>.GSW>
     {
-        internal ValueTypeDefaultComparerFrozenSet(HashSet<T> source) : base(source)
+        internal ValueTypeDefaultComparerFrozenSet(HashSet<T> source) : base(source, Constants.KeysAreHashCodes<T>())
         {
             Debug.Assert(typeof(T).IsValueType);
             Debug.Assert(ReferenceEquals(source.Comparer, EqualityComparer<T>.Default));