- don't use custom ToArray for small frozen collections, up to 50% gain for creation time for collections with <= 4 items

|                            Type |             Method |        Job | Count |        Mean | Ratio | Allocated | Alloc Ratio |
|-------------------------------- |------------------- |----------- |------ |------------:|------:|----------:|------------:|
|      Perf_Frozen&lt;ReferenceType&gt; | ToFrozenDictionary |         PR |     4 |    98.70 ns |  0.54 |     152 B |        0.66 |
|      Perf_Frozen&lt;ReferenceType&gt; | ToFrozenDictionary |       main |     4 |   182.50 ns |  1.00 |     232 B |        1.00 |
|                                 |                    |            |       |             |       |           |             |
|      Perf_Frozen&lt;ReferenceType&gt; |        ToFrozenSet |         PR |     4 |   165.18 ns |  0.83 |     384 B |        0.91 |
|      Perf_Frozen&lt;ReferenceType&gt; |        ToFrozenSet |       main |     4 |   198.46 ns |  1.00 |     424 B |        1.00 |

- use the int trick for uint, short, ushort, sbyte and byte: we know that all hash codes are unique and we can avoid some work later. 10-15% CPU time gain and 15-20% allocation reduction for FrozenDictionary and FrozenHashSet where TKey is uint, short, ushort, byte, sbyte

|             Method |  Job | Count |         Mean | Ratio | Allocated | Alloc Ratio |
|------------------- |----- |------ |-------------:|------:|----------:|------------:|
| ToFrozenDictionary |   PR |    64 |  2,808.38 ns |  0.91 |    4400 B |        0.78 |
| ToFrozenDictionary | main |    64 |  3,072.64 ns |  1.00 |    5656 B |        1.00 |
|                    |      |       |              |       |           |             |
|        ToFrozenSet |   PR |    64 |  2,964.31 ns |  0.87 |    5408 B |        0.81 |
|        ToFrozenSet | main |    64 |  3,420.82 ns |  1.00 |    6664 B |        1.00 |
|                    |      |       |              |       |           |             |
| ToFrozenDictionary |   PR |   512 | 25,362.06 ns |  0.87 |   45216 B |        0.84 |
| ToFrozenDictionary | main |   512 | 28,944.59 ns |  1.00 |   53672 B |        1.00 |
|                    |      |       |              |       |           |             |
|        ToFrozenSet |   PR |   512 | 27,882.97 ns |  0.84 |   51632 B |        0.86 |
|        ToFrozenSet | main |   512 | 33,151.36 ns |  1.00 |   60088 B |        1.00 |

- move Length Buckets code to a dedicated helper type to reduce code duplication and decrease code size

@stephentoub this is my last PR for frozen collections perf for now

I apologize if the CI gets red, for some reason I am getting a build error and I temporarily can't run unit tests locally:

```log
CSC : error CS8785: Generator 'JSExportGenerator' failed to generate source. It will not contribute to the output a
nd compilation errors may occur as a result. Exception was of type 'TypeLoadException' with message 'Could not load
 type 'Microsoft.Interop.IGeneratorDiagnostics' from assembly 'Microsoft.Interop.SourceGeneration, Version=42.42.42
.42, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a'.' [D:\projects\runtime\src\libraries\System.Collections.Immu
table\src\System.Collections.Immutable.csproj::TargetFramework=net7.0]
CSC : error CS8785: Generator 'JSImportGenerator' failed to generate source. It will not contribute to the output a
nd compilation errors may occur as a result. Exception was of type 'TypeLoadException' with message 'Could not load
 type 'Microsoft.Interop.IGeneratorDiagnostics' from assembly 'Microsoft.Interop.SourceGeneration, Version=42.42.42
.42, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a'.' [D:\projects\runtime\src\libraries\System.Collections.Immu
table\src\System.Collections.Immutable.csproj::TargetFramework=net7.0]
```

contributes to #87964 and closes #87964 it as it's the last PR in the series