@adamsitnik 2023-06-28
>  I believe I fixed the build break you're seeing, so I want to make sure it didn't break in a different way.

@jkoritzinsky I've synced my branch with upstream and everything works fine. Thanks for letting me know!